Don't use the constructor call trick when a custom element is constructed with new
We need to use the constructor call trick in the PivotCtor when the element has been created by the browser directly (e.g. document.createElement, or via the parser) because in that case the only constructor that's called is the PivotCtor.

However, if the element is called via `new MyElement()` then the MyElement constructor (and any superclass constructors) will be called properly without any work on our part.

This change fixes a bug we noticed where calling `new MyElement()` was running constructors twice.