Use a single definition per tag name
Update it when the tag is redefined.

This way LatestCtor and the callbacks are updated for all instances immediately on redefinition.

I'm not certain this change is an improvement in all cases, but it's the behavior that I think we generally want when doing HMR in google, but I could see the value in making it an option.